Switch to using global state and reducer
